requirements.txt
streamlit
pandas
newsapi-python
google-generativeai
python-dotenv
